+++ source/Common/hexint.h
@@ -4,9 +4,9 @@
 #include <QString>
 
 namespace HexInt{
-    QString intToString(const quint16 value, const bool prefix = true, const qsizetype precision = 4);
-    quint16 hexStringToInt(const QString &string, bool* okptr);
-    quint16 stringToInt(const QString &string, bool* okptr);
+    QString intToString(const quint32 value, const bool prefix = true, const qsizetype precision = 4);
+    quint32 hexStringToInt(const QString &string, bool* okptr);
+    quint32 stringToInt(const QString &string, bool* okptr);
 };
 
 #endif // HEXINT_H
